+++ ParticleSimulation/src/particlesimulation/Power.java
@@ -1,17 +1,7 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
 package particlesimulation;
 
-/**
- *
- * @author Robbykunsan
- */
 public class Power extends DirectionVector{
-	public Power(double x, double y, double z){
+	public Power(float x, float y, float z){
 		super(x, y, z);
 	}
